pipico/read_temp_data/plot/requirements.txt
pyserial==3.5
matplotlib==3.10.8